backend/requirements.txt
Flask==3.0.0
Flask-Session==0.5.0
psycopg2-binary>=2.9.9,<3.0.0
pytz
requests
gunicorn
openpyxl
python-dotenv